Windsurf/requirements.txt
zscaler-sdk-python
python-dotenv>=1.0.0
